==> 1_assemble.Rmd <==
---
title: "1_Assemble"
author: "Helena"
date: '2023-03-09'
output:
  github_document: default
  word_document:
    reference_docx: word_style1.docx
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Assembly
This markdown file explains the code in 1_assemble.R. The raw dataset is a series of csv files, one csv file for each year (2007-2016). The first step is to combine these sets into one big table. Like most data which has been manually collated in excel there is inconsistency among the column names which makes this tricky... The following script works reproducibly with the csv files used for the paper but may need tweaking to be effective on other files. I could have just manually fixed the problems with the csvs but I thought this would be a better way of doing it if you had loads of csv files. 

```{r message = FALSE, warning = FALSE}
#Load required packages
library(tidyverse)
library(here)
```

A quirk of the csvs is some of them have an extra column, which when reading the files in in the loop below, causes the datasets without it to have a blank column called '27', which causes problems when you try and combine the datasets, so I wrote a function to deal with this if it is present.

```{r}
#This function removes column 27 if present (otherise causes problems) -----
cond_remove <- function(df) {
  colnames <- colnames(df)
  if (any(str_detect(colnames, "27"))) {
    df <- df %>% select(-"27")
  } else {
    df <- df
  }
  return(df)
}
```

This makes a list of the data file names:

```{r}
files <- list.files(here("data", "csvs"))
```

Make a blank tibble imaginatively called df.

```{r}
df <- tibble()
```

This loop reads in the data, one file at a time, into a table called new_df. It then selects the correct columns and alters the names of columns to make sure they are consistent across files (if you are doing this with different files this will take some experimentation), makes all column names lower case, applies the function to get rid of 27 where it is present and lastly add all the rows to the blank table created in the last step.

```{r, warning=FALSE, message=FALSE, error=FALSE}
for (i in seq_along(files)) {
  new_df <- read_csv(str_c("data/csvs/", files[i])) %>%
    select(1:27) %>%
    rename_with(~str_replace_all(., "\n|/|\\(mm\\)|cells|ID|No|\\.| ", ""),
                everything()) %>%
    rename_with(~str_to_lower(.)) %>%
    cond_remove(.) %>%
    mutate(across(where(is.numeric), ~as.character(.)))

  df <- bind_rows(df, new_df)
}
```

Create a rowid and save the dataset as assembled_data.RDS

```{r}
df %>%
  rowid_to_column(var = "rowid") %>%
  saveRDS(here("data", "assembled_data.RDS"))
```

The saved dataset should look like this (the following data is all made up):

```{r, include= FALSE}
fake_ad <- read_csv(here("data", "fake_ad.csv")) %>%
  mutate(across(where(is.numeric), ~as.character(.)))
```

```{r}
glimpse(fake_ad)
```
